Added more self = [super init] calls.  added more NSInteger datatypes, replaced a couple deprecreated method calls, fixed some memory leaks.  minor stuff.

+++ EOControl/Global ID/EOKeyGlobalID.h
@@ -30,11 +30,11 @@ http://www.raftis.net/~alex/
 
 @interface EOKeyGlobalID : EOGlobalID
 {
-   NSString			*entityName;
-   NSString			**keys;
-   id					*values;
-   int				count;
-	unsigned int	hash;
+    NSString		*entityName;
+    NSString		**keys;
+    id				*values;
+    int				count;
+	NSUInteger      hash;
 }
 
 + (id)globalIDWithEntityName:(NSString *)anEntityName keys:(NSString **)primaryKeys values:(id *)someValues count:(int)aCount;
